final printf

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: anmindt <anmindt@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/05 16:36:50 by anmindt           #+#    #+#             */
-/*   Updated: 2024/03/08 19:20:25 by anmindt          ###   ########.fr       */
+/*   Updated: 2024/03/11 14:19:46 by anmindt          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,7 @@
 # include <stdio.h>
 # include <stdarg.h>
 # include <limits.h>
+# include <stdint.h>
 
 //core functions for initialising printing with specific format
 int		ft_printf(const char *str, ...);
@@ -27,10 +28,10 @@ void	print_form(char str, int *count, va_list argptr);
 //functions for converting and printing numbers
 void	ft_print_nbr(int n, int *count);
 void	ft_print_unum(unsigned int n, int *count);
-int		ft_narrlen(unsigned int n);
-char	*ft_narr(unsigned int n);
+int		ft_narrlen(unsigned long long n);
+char	*ft_narr(unsigned long long n);
 void	ft_print_and_free(char *str, int *count);
-void	ft_dectohex_up(unsigned int n, char c, int *count);
+void	ft_dectohex_up(unsigned long long n, char c, int *count);
 
 //functions for printing chars and strings
 void	print_c(char c, int *count);
